html/semantics/scripting-1/the-script-element/module/inline-async-inserted-execorder.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Inline async="" module scripts execute or throw parse errors asynchronously</title>
<link rel="help" href="https://github.com/whatwg/html/issues/9864">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<script>
"use strict";
setup({ allow_uncaught_exception: true });

async_test(t => {
  window.log = ["before any script execution"];

  window.addEventListener("error", ev => {
    window.log.push("error event on Window");
  });

  const noErrorScript = document.createElement("script");
  noErrorScript.type = "module";
  noErrorScript.async = true;
  noErrorScript.textContent = "window.log.push('no error script executed');";

  // This should queue a microtask to run the script, but not run it immediately.
  document.head.append(noErrorScript);

  log.push("after inserting noErrorScript");
  assert_array_equals(window.log, [
    "before any script execution",
    "after inserting noErrorScript"
  ]);

  const parseErrorScript = document.createElement("script");
  parseErrorScript.type = "module";
  parseErrorScript.async = true;
  parseErrorScript.textContent = "+++++";

  // This should queue a microtask to fire the error event, but not fire it immediately.
  document.head.append(parseErrorScript);

  log.push("after inserting parseErrorScript");
  assert_array_equals(window.log, [
    "before any script execution",
    "after inserting noErrorScript",
    "after inserting parseErrorScript"
  ]);

  queueMicrotask(t.step_func_done(() => {
    assert_array_equals(window.log, [
      "before any script execution",
      "after inserting noErrorScript",
      "after inserting parseErrorScript",
      "no error script executed",
      "error event on Window"
    ]);
  }));
});
</script>
